Added some networking code.
GamePooledByteBufferAllocator is a wrapper that delegates to the PooledByteBufferAllocator.Default instance.

The purpose of this is to add some front-padding on buffer creation, which gives us some wiggle-room that will hopefully allow us to avoid creating a new buffer when processing packets with a 2-byte data length in GameMessageDecoder.

Note that GameMessageDecoder is a work-in-progress.

Further, I'm hoping to do without the Packet class. Instead, I'd rather use extension methods on IByteBuffer, thereby avoiding the allocation of new Packet objects.