Added the concept of thread guards. Thread guards are a somewhat borrowed idea from DotNetty.
DotNetty checks the current thread when attempting to queue work for an event loop. If the current thread is the event loop thread then it just queues it directly; otherwise, it does it in a roundabout but thread-safe manner.

Thread guards in OpenClassic are designed to ensure that game content runs exclusively on the game thread, that networking code does NOT run on the game thread, and that other stuff (such as IO-heavy work) also does NOT run on the game thread.